+++ lib/reconcile/cfssl_ca_secret.go
@@ -23,7 +23,7 @@ package reconcile
 import (
 	"context"
 
-	apiv1alpha1 "github.com/astarte-platform/astarte-kubernetes-operator/pkg/apis/api/v1alpha1"
+	apiv1alpha1 "github.com/astarte-platform/astarte-kubernetes-operator/api/v1alpha1"
 	"github.com/openlyinc/pointy"
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"