+++ test/utils/test_delete_astarte.go
@@ -2,12 +2,10 @@ package utils
 
 import (
 	goctx "context"
-	"fmt"
 
-	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	operator "github.com/astarte-platform/astarte-kubernetes-operator/pkg/apis/api/v1alpha1"
+	operator "github.com/astarte-platform/astarte-kubernetes-operator/api/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
@@ -15,26 +13,22 @@ import (
 )
 
 // AstarteDeleteTest deletes an Astarte instance and tests whether it was deleted and cleaned up
-func AstarteDeleteTest(f *framework.Framework, ctx *framework.Context) error {
-	namespace, err := ctx.GetWatchNamespace()
-	if err != nil {
-		return fmt.Errorf("could not get namespace: %v", err)
-	}
+func AstarteDeleteTest(c client.Client, namespace string) error {
 	installedAstarte := &operator.Astarte{}
 	// use Context's helper to Get the object
-	if err := f.Client.Get(goctx.TODO(), types.NamespacedName{Name: AstarteTestResource.GetName(), Namespace: namespace}, installedAstarte); err != nil {
+	if err := c.Get(goctx.TODO(), types.NamespacedName{Name: AstarteTestResource.GetName(), Namespace: namespace}, installedAstarte); err != nil {
 		return err
 	}
 
 	// Delete the object
-	if err := f.Client.Delete(goctx.TODO(), installedAstarte); err != nil {
+	if err := c.Delete(goctx.TODO(), installedAstarte); err != nil {
 		return err
 	}
 
 	// Wait until everything in the namespace is erased. Finalizers should do the job.
 	if err := wait.Poll(DefaultRetryInterval, DefaultTimeout, func() (done bool, err error) {
 		deployments := &appsv1.DeploymentList{}
-		if err = f.Client.List(goctx.TODO(), deployments, client.InNamespace(namespace)); err != nil {
+		if err = c.List(goctx.TODO(), deployments, client.InNamespace(namespace)); err != nil {
 			return false, err
 		}
 		if len(deployments.Items) > 0 {
@@ -42,7 +36,7 @@ func AstarteDeleteTest(f *framework.Framework, ctx *framework.Context) error {
 		}
 
 		statefulSets := &appsv1.StatefulSetList{}
-		if err = f.Client.List(goctx.TODO(), statefulSets, client.InNamespace(namespace)); err != nil {
+		if err = c.List(goctx.TODO(), statefulSets, client.InNamespace(namespace)); err != nil {
 			return false, err
 		}
 		if len(statefulSets.Items) > 0 {
@@ -50,7 +44,7 @@ func AstarteDeleteTest(f *framework.Framework, ctx *framework.Context) error {
 		}
 
 		configMaps := &v1.ConfigMapList{}
-		if err = f.Client.List(goctx.TODO(), configMaps, client.InNamespace(namespace)); err != nil {
+		if err = c.List(goctx.TODO(), configMaps, client.InNamespace(namespace)); err != nil {
 			return false, err
 		}
 		if len(configMaps.Items) > 0 {
@@ -58,7 +52,7 @@ func AstarteDeleteTest(f *framework.Framework, ctx *framework.Context) error {
 		}
 
 		secrets := &v1.SecretList{}
-		if err = f.Client.List(goctx.TODO(), secrets, client.InNamespace(namespace)); err != nil {
+		if err = c.List(goctx.TODO(), secrets, client.InNamespace(namespace)); err != nil {
 			return false, err
 		}
 		// The Default Token is acceptable.
@@ -67,7 +61,7 @@ func AstarteDeleteTest(f *framework.Framework, ctx *framework.Context) error {
 		}
 
 		pvcs := &v1.PersistentVolumeClaimList{}
-		if err = f.Client.List(goctx.TODO(), pvcs, client.InNamespace(namespace)); err != nil {
+		if err = c.List(goctx.TODO(), pvcs, client.InNamespace(namespace)); err != nil {
 			return false, err
 		}
 		if len(pvcs.Items) > 0 {